requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
configparser==7.1.0
contourpy==1.1.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.14.1
zipp==3.20.2